PL-201 - ctheme -> mtheme, удалены не нужные комменты, вынесены classnames из render
