+++ src/assets/css/typo/typo.less
@@ -104,16 +104,16 @@
 }
 
 // TODO: для теста темной темы вк
-.typo .ctheme--darkpic a {
+.typo .mtheme--darkpic a {
   color: @c-white;
 }
 
-.ctheme--darkpic {
+.mtheme--darkpic {
   .typo-links {
     color: @c-white;
   }
 }
 
-.typo-links .ctheme--darkpic a {
+.typo-links .mtheme--darkpic a {
   .tpui-mx-color-links-full-base(@c-white, @c-primary-blue)
 }
