+++ src/components/forms/Checkbox/ccheckbox.less
@@ -90,7 +90,7 @@
     }
   }
 
-  .ctheme--lightpic & {
+  .mtheme--lightpic & {
     .ccheckbox__icon {
       border-color: @c-gray-50-fade-35;
     }
@@ -112,7 +112,7 @@
     }
   }
 
-  .ctheme--darkpic & {
+  .mtheme--darkpic & {
     .ccheckbox__icon {
       border-color: @c-gray-50-fade-35;
     }
