+++ src/components/forms/Radio/cradio.less
@@ -88,7 +88,7 @@
     }
   }
 
-  .ctheme--lightpic & {
+  .mtheme--lightpic & {
     .cradio__icon {
       border-color: @c-gray-50-fade-35;
     }
@@ -110,7 +110,7 @@
     }
   }
 
-  .ctheme--darkpic & {
+  .mtheme--darkpic & {
     .cradio__icon {
       border-color: @c-gray-50-fade-35;
     }
