+++ src/components/forms/RadioButton/cradiobutton.less
@@ -119,7 +119,7 @@
 
   }
 
-  .ctheme--lightpic & {
+  .mtheme--lightpic & {
     --icon-color: @c-black;
     --border-color: @c-gray-50-fade-35;
 
@@ -139,7 +139,7 @@
     }
   }
 
-  .ctheme--darkpic & {
+  .mtheme--darkpic & {
     --icon-color: @c-white;
     --border-color: @c-gray-50-fade-35;
